+++ extensions/zendesk/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Zendesk Changelog
 
-## [Add Action to Open Article in Browser] - {PR_MERGE_DATE}
+## [Fetch More Than 25 Articles via Pagination] - {PR_MERGE_DATE}
 
-- The `Open in Browser` action goes to the JSON article. There is now a new Action with shortcut `Cmd + O` that goes to the article instead.
+- Paginate the query
 - Some refactoring of the code to use the `useFetch` hook
 
 ## [Updates] - 2022-12-22